requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
Farama-Notifications==0.0.4
flatbuffers==24.12.23
fonttools==4.55.3
gast==0.6.0
gin-config==0.5.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.68.1
gym==0.23.1
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.12.1
idna==3.10
keras==3.7.0
keras-nightly==3.7.0.dev2024123003
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.2.1
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf-agents-nightly==0.20.0.dev20231221
tf-nightly==2.19.0.dev20241219
tf_keras-nightly==2.19.0.dev2024123010
tf_nightly_intel==2.19.0.dev20241219
tfp-nightly==0.26.0.dev20241230
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
